Make client fixtures available to all tests
