+++ tests/integration/conftest.py
@@ -9,7 +9,6 @@ import multiprocessing as mp
 import os
 import threading
 
-import boto3
 import pytest
 
 from localstack import config
@@ -17,12 +16,6 @@ from localstack.config import is_env_true
 from localstack.constants import ENV_INTERNAL_TEST_RUN
 from localstack.runtime import events
 from localstack.services import infra
-from localstack.testing.aws.util import (
-    base_aws_client_factory,
-    base_aws_session,
-    primary_testing_aws_client,
-    secondary_testing_aws_client,
-)
 from localstack.utils.common import safe_requests
 from tests.integration.test_es import install_async as es_install_async
 from tests.integration.test_opensearch import install_async as opensearch_install_async
@@ -183,35 +176,3 @@ def localstack_runtime():
     localstack_started.wait()
     yield
     return
-
-
-@pytest.fixture(scope="session")
-def aws_session() -> boto3.Session:
-    """
-    Returns a boto3 session.
-    """
-    return base_aws_session()
-
-
-@pytest.fixture(scope="session")
-def aws_client_factory(aws_session):
-    """
-    Returns a client factory for testing.
-    """
-    return base_aws_client_factory(aws_session)
-
-
-@pytest.fixture(scope="session")
-def aws_client(aws_client_factory):
-    """
-    Returns an AWS client configured with primary test credentials.
-    """
-    return primary_testing_aws_client(aws_client_factory)
-
-
-@pytest.fixture(scope="session")
-def secondary_aws_client(aws_client_factory):
-    """
-    Returns an AWS client configured with secondary test credentials.
-    """
-    return secondary_testing_aws_client(aws_client_factory)
